refactor(slt_common): OdomDataBuffer uses int64_t as map key
- Internal map now stores timestamps as int64_t (microseconds)
- Interface unchanged (double time parameter)
- get_interpolated_data returns exact match directly without interpolation

+++ slt_common/include/slt_common/odom_data_buffer.hpp
@@ -44,7 +44,7 @@ public:
   void set_max_buffer_size(size_t max_buffer_size);
 
 private:
-  std::map<double, slt_common::OdomData> buffer_;
+  std::map<int64_t, slt_common::OdomData> buffer_;
   size_t max_buffer_size_;
 };
 }  // namespace slt_common
